libnm: use plain string comparison for mac-address properties
Historically, mac-address properties were compared by converting the
string to a "ay" GVariant. This way, it could not represent invalid
strings and would be case-insensitive.

Now all property setters normalize the strings first, with
_nm_utils_hwaddr_canonical_or_invalid(). So usually, we would not expect
different cases.

Drop the special compare function and do plain string comparison. This
seems better, because there is no extra code and it's also more accurate
to the real property (in the GObject instance).

There is a subtle differences: the normalization only normalizes strings of
the expected length. But the previous approach would compare strings of any
length. But wrong lengths are also considered invalid, and we don't guarantee
proper comparison for invalid settings.