libnm: fix clearing parentheses in nm_utils_fixup_product_string()
Previously, once in_parent was TRUE it was never reset, thus the
remainder of the string was cleared. That was most likely not intended.

If the intent really was to clear all the remainder, then the code could
have simply truncated the string at the first '('.

(cherry picked from commit 17bdd3a40d7bc0bbba52ad333be4b3210fa3981a)