std-aux: implement NM_PTRARRAY_LEN() macro via static function
We use NM_PTRARRAY_LEN(), and I find it a bit ugly that a macro does so
much. Maybe, it's better to have it as a function.

But the macro currently lives in "libnm-std-aux/nm-std-aux.h", which
is header-only. To add it to a C source file, we would have to move
it to another header, but "libnm-std-aux/nm-std-aux.h" is nice because
it gets included by default already.

Keep it in "libnm-std-aux/nm-std-aux.h", but implement it as an inline
function.

The macro now only does (as before) some type checking shenanigans to ensure
that the argument is a pointer to pointers.

In practice, there is probably very little difference compared to
the macro before, likely the code will anyway be inlined.