nm-sudo: don't register pending job for current operations
Currently we only implmement two operations (Ping() and GetFD()). Both
complete right away. There is no need to register a pending job, if
the job does not get processed asynchronously.

In the future, we may have methods that need asynchronous processing
and where we need to register them as pending job.