meson: Allow building only libnm
The Steam Runtime includes a 32-bit libnm.so. We would like to be able
to build only libnm from NetworkManager and minimize the number of
32-bit dependencies required to do so.

There is not explicit support for this in the autotools build system,
but it was still possible to accomplish with

  make src/libnm-client-impl/libnm.la
  make \
    install-libLTLIBRARIES
    install-libnmincludeHEADERS
    install-nodist_libnmincludeHEADERS
    install-pkgconfigDATA

Unfortunately ninja does not provide a way to install only specific
targets.